Prospective config refactor into modules
