+++ pgdog/src/config/auth.rs
@@ -1,0 +1,47 @@
+use serde::{Deserialize, Serialize};
+use std::str::FromStr;
+
+#[derive(Serialize, Deserialize, Debug, Clone, Default, PartialEq)]
+#[serde(rename_all = "snake_case")]
+pub enum PassthoughAuth {
+    #[default]
+    Disabled,
+    Enabled,
+    EnabledPlain,
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone, Default, PartialEq)]
+#[serde(rename_all = "snake_case")]
+pub enum AuthType {
+    Md5,
+    #[default]
+    Scram,
+    Trust,
+}
+
+impl AuthType {
+    pub fn md5(&self) -> bool {
+        matches!(self, Self::Md5)
+    }
+
+    pub fn scram(&self) -> bool {
+        matches!(self, Self::Scram)
+    }
+
+    pub fn trust(&self) -> bool {
+        matches!(self, Self::Trust)
+    }
+}
+
+impl FromStr for AuthType {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_str() {
+            "md5" => Ok(Self::Md5),
+            "scram" => Ok(Self::Scram),
+            "trust" => Ok(Self::Trust),
+            _ => Err(format!("Invalid auth type: {}", s)),
+        }
+    }
+}
